requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.12.1
asttokens==3.0.0
async-timeout==5.0.1
attributedict==0.3.0
attrs==25.3.0
backports.cached-property==1.0.2
blessings==1.7
cachetools==6.1.0
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
contourpy==1.3.2
coverage==7.9.2
cycler==0.12.1
decorator==5.2.1
deepdiff==8.5.0
distlib==0.4.0
easydict==1.13
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.73.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.1
humanfriendly==10.0
idna==3.10
inspecta==0.1.3
iopath==0.1.10
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.15.0
llvmlite==0.45.1
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
orderly-set==5.5.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
pyproject-api==1.9.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
PyYAML==6.0.2
requests==2.32.4
rootpath==0.1.1
safetensors==0.7.0
scikit-learn==1.7.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.24
tomli==2.2.1
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchmetrics==1.8.0
torchsparse==2.1.0
torchvision==0.19.0+cu124
tox==4.28.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
typer-slim==0.21.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.32.0
wcwidth==0.2.14
Werkzeug==3.1.3
yacs==0.1.8
yarl==1.20.1
